+++ packages/forms/src/index.ts
@@ -1,3 +1,4 @@
+export { errorMessage, FORM_ERROR_KEY, serverErrorToFieldErrors } from "./server-errors.js";
 export {
   useForm,
   zodErrorsToFieldErrors,